More logging to determine the issue and TODOs/FIXMEs.
To avoid invocation on each Init(), the re-balancing should
likely be moved to Session and done there through the
definition manager.

This also paves the way for a more generic Rebalancer -
once a few "examples" are done here for the HE/OG, an
API or config syntax is easier to devise.