Config: update to match pre-1.197.
From https://www.reddit.com/r/spaceengineers/comments/k23hcz/update_v197_hydrogen_rebalance/:

> Hydrogen Engines only consume half as much hydrogen (per watt produced) now.
>
> O2/H2 Generators now use 1/2 the amount of ice per second (and produce twice as much hydrogen per unit of ice).

It seems that the changes can be summarised in mod-config terms as a simple:

<HydrogenEngineEfficiencyMultiplier>PREVIOUS*2</HydrogenEngineEfficiencyMultiplier>

This can be confirmed in data files: the vanilla FuelProductionToCapacityMultiplier for hydrogen engines has been bumped from 0.005 to 0.01 (doubled).

For that reason, the mod's HydrogenEngineEfficiencyMultiplier is halved.