Config, misc: further reduce O2/H2 gen speed to prevent "recharging" batteries with ice.
The second change:

> O2/H2 Generators now use 1/2 the amount of ice per second (and produce twice as much hydrogen per unit of ice).

... has had a respective change in the data file, too:
IceConsumptionPerSecond was reduced from 50 to 25
in vanilla.

Strangely enough, this did not soften the physics
violation. In fact, it has again become possible to go the
(ice) -> O2/H2 gen -> H2 engine -> battery route to
produce surlpus electricity. :/

In total, it seems the route has been buffed 4 times.
Previous commits de-buffed it 2 times; this de-buffs it
2 more.